Merge conflict with latest moses

Conflicts:
	OnDiskPt/Main.cpp
	contrib/mert-moses-multi.pl
	contrib/other-builds/CreateOnDisk.xcodeproj/project.pbxproj
	contrib/other-builds/moses-chart-cmd.xcodeproj/project.pbxproj
	contrib/other-builds/moses-cmd.xcodeproj/project.pbxproj
	contrib/other-builds/moses.xcodeproj/project.pbxproj
	lm/model_test.cc
	lm/search_trie.hh
	mert/Point.cpp
	mert/extractor.cpp
	mert/mert.cpp
	mert/test_scorer.cpp
	moses/LM/IRST.cpp
	moses/LM/Remote.h
	moses/ThreadPool.h
	moses/TranslationModel/BilingualDynSuffixArray.cpp
	moses/TranslationModel/DynSAInclude/FileHandler.cpp
	moses/TranslationModel/DynSAInclude/types.h
	scripts/ems/experiment.meta
	scripts/ems/experiment.perl
	scripts/training/train-model.perl
	util/probing_hash_table_test.cc

+++ misc/queryLexicalTable.cpp
@@ -1,11 +1,11 @@
 #include <iostream>
 #include <string>
 
-#include "Phrase.h"
-#include "FactorCollection.h"
-#include "Timer.h"
-#include "InputFileStream.h"
-#include "LexicalReorderingTable.h"
+#include "moses/Phrase.h"
+#include "moses/FactorCollection.h"
+#include "moses/Timer.h"
+#include "moses/InputFileStream.h"
+#include "moses/LexicalReorderingTable.h"
 
 using namespace Moses;
 
@@ -24,7 +24,7 @@ void printHelp()
 
 std::ostream& operator<<(std::ostream& o, Scores s)
 {
-  for(int i = 0; i < s.size(); ++i) {
+  for(size_t i = 0; i < s.size(); ++i) {
     o << s[i] << "  ";
   }
   //o << std::endln;
